requirements.txt
Flask
python-docx
requests
